Do not depend on qgis_plugin_tools if it is not chosen
